changed the delete_automatically func

+++ SmartRm/Trash.py
@@ -10,27 +10,33 @@ from datetime import datetime
 
 # new procedure if trash does not exist ( create new trash)
 
-class Trash():
-    def __init__(self, path_trash, path_log_j, path_log_t, p, s, cap, t):
+class Trash(object):
+    def __init__(self, path_trash, path_log_j, path_log_t, policy_time, policy_size, size, cur_size, capacity, time):
         self.path = path_trash
         self.log_writer = Logwriter.Logwriter(path_log_j, path_log_t)
-        self.politics = p
-        self.size = s
-        self.capacity = cap
-        self.time = t
+        self.policy_time = policy_time
+        self.policy_size = policy_size
+        self.max_size = size
+        self.cur_size = cur_size
+        self.max_capacity = capacity  # the max quantity of files in trash
+        self.max_time = time
 
-    def delete_automatically(self):  # not checked(means clean trash)
+    def delete_automatically(self):  # works
         # delete the whole trash
-        # os.access  - check the access!!!!(права доступа)
         d = os.listdir(self.path)
         for item in d:
             subpath = os.path.join(self.path, item)  # формирование адреса
-            if item == 'Trash_log':
-                continue
-            elif os.path.isdir(subpath):
-                shutil.rmtree(subpath)
-            elif not os.path.isdir(subpath):
-                os.remove(subpath)
+            dict_contains = False
+            for _dict in self.log_writer.file_dict_arr:
+                if _dict['id'] == item:
+                    dict_contains = True
+                    break
+            if dict_contains:
+                if os.path.isdir(subpath):
+                    shutil.rmtree(subpath)
+                elif not os.path.isdir(subpath):
+                    os.remove(subpath)
+
         clean_json = open(self.log_writer.file_dict_path, 'w')
         clean_json.close()
         clean_txt = open(self.log_writer.file_dict_path_txt, 'w')
@@ -52,11 +58,6 @@ class Trash():
             subpath = os.path.join(path, item)  # формирование адреса
             if file_id in subpath:
                 return subpath
-            # else:
-            #     if os.path.isdir(subpath):
-            #         return self.get_path_by_id(file_id, subpath)
-            #     elif not os.path.isdir(subpath):
-            #         continue
 
     def watch_trash(self):  # not checked
         if self.log_writer.file_dict_arr is [] or self.log_writer.file_dict_arr is None or self.log_writer.file_dict_arr.__len__() == 0:
@@ -64,19 +65,10 @@ class Trash():
         else:
             txt_file = open(self.log_writer.file_dict_path_txt, 'r')
             print(txt_file.read())
-            # for line in txt_file:
-            #     print line
-
-        # file_list = json.load(open(self.trash_log_path, 'r'))
-        # if not file_list:
-        #     print 'trash bucket is empty'
-        # else:
-        #     for item in file_list:  # check this
-        #         print item['name']  # check this
-        #         # file_list = os.listdir(self.trash_path)
 
     def restore_trash_automatically(self):  # not done
         # restore the the whole trash
+        # the same strategy as delete
         self.rename_all_directory_content(self.path)
         allfiles = os.listdir(self.path)
         for f in allfiles:
@@ -84,32 +76,21 @@ class Trash():
                 new_name = self.log_writer.get_name(f)
         return None
 
-    def rename_all_directory_content(self, path):  # not checked
-
-        if not os.path.isdir(path):
-            name = os.path.split(path)
-            newname = self.log_writer.get_name(name[1])
-            os.rename(path, name[0]+newname)
-        elif os.path.isdir(path):
-            index = 0
-            for i in reversed(range(len(path))):
-                if path[i] == '/':
-                    index = i
-                    break
-            dirname = path[index + 1:]
-            d = os.listdir(path)
-            # print d
-            # for item in d:
-            #     subpath = os.path.join(path, item)  # формирование адреса
-            #     if os.path.isdir(subpath):
-            #         self.rename_all_directory_content(subpath)
-            #         #subfile = self.write_file_dict(subpath)
-            #
-            #     elif not os.path.isdir(subpath):
-            #         self.rename_all_directory_content(subpath)
-            #         #subdict = self.write_file_dict(subpath)
-            newname = self.log_writer.get_name(dirname)
-            os.rename(path, path[:path.__sizeof__() - (index+1)]+newname)
+    # def rename_all_directory_content(self, path):  # not checked
+    #
+    #     if not os.path.isdir(path):
+    #         name = os.path.split(path)
+    #         new_name = self.log_writer.get_name(name[1])
+    #         os.rename(path, name[0]+new_name)
+    #     elif os.path.isdir(path):
+    #         index = 0
+    #         for i in reversed(range(len(path))):
+    #             if path[i] == '/':
+    #                 index = i
+    #                 break
+    #         dir_name = path[index + 1:]
+    #         new_name = self.log_writer.get_name(dir_name)
+    #         os.rename(path, path[:path.__sizeof__() - (index+1)]+new_name)
 
     def restore_trash_manually(self, path):  # not checked
         # restore one file in the trash
@@ -132,27 +113,33 @@ class Trash():
         self.log_writer.write_to_txt()
         return None
 
-    def check_politics(self):
-        # config = json.load(open('SmartRm/Configure.json', 'r'))
-        # politics = config['politics']
-        if politics == 'time':
-            print 'time'
-        elif politics == 'size':
+    def check_policy(self, path):  # not checked(redo to check the whole bucket)
+        if self.policy_time:
+            confirm = self.check_date_if_overflow(path)
+            if confirm:
+                self.delete_manually(path)
+        elif self.policy_size:
             print 'size'
-        elif politics == 'combined':
-            print 'combined'
 
-    def check_date(self, path):
+    def count_days(self, path):
         name = os.path.split(path)
         if name[1] != '':
             file_id = self.log_writer.get_id(name[1])
             file_date = self.log_writer.get_date(file_id)
-            cur_date = datetime.strftime(datetime.now(), '%Y.%m.%d %H:%M:%S')
-            years = ((int)(file_date[0]+file_date[1]) - (int)(cur_date[0]+cur_date[1]))
-            months = ((int)(file_date[4] + file_date[5]) - (int)(cur_date[4] + cur_date[5]))
-            days = ((int)(file_date[7] + file_date[8]) - (int)(cur_date[7] + cur_date[8]))
-            days_common = years*365+months*30+days
-        return days_common
+            cur_date = datetime.date.today()
+            file_date = file_date.split('-')
+            date_a = datetime.date(int(file_date[0]), int(file_date[1]), int(file_date[2]))
+
+            days = cur_date - date_a
+            days = str(days).split()[0]
+        return days
+
+    def check_date_if_overflow(self, path):
+        time = self.count_days(path)
+        if time > self.max_time:
+            return True
+        else:
+            return False
 
     def check_size(self):
         return None
